requirements.txt
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv